+++ Libplanet.Explorer/GraphTypes/TransactionType.cs
@@ -17,25 +17,52 @@ namespace Libplanet.Explorer.GraphTypes
         public TransactionType()
         {
             Field<NonNullGraphType<IdGraphType>>(
-                "Id",
-                resolve: ctx => ctx.Source.Id.ToString());
-            Field(x => x.Nonce);
-            Field(x => x.Signer, type: typeof(NonNullGraphType<AddressType>));
-            Field<NonNullGraphType<ByteStringType>>(
-                "PublicKey",
-                resolve: ctx => ctx.Source.PublicKey.Format(true)
+            name: "Id",
+            description: "A unique identifier derived from this transaction content.",
+            resolve: ctx => ctx.Source.Id.ToString());
+            Field<NonNullGraphType<LongGraphType>>(
+            name: "Nonce",
+            description: "The number of previous transactions committed by the signer of this tx.",
+            resolve: x => x.Source.Nonce
             );
             Field(
-                x => x.UpdatedAddresses,
-                type: typeof(NonNullGraphType<ListGraphType<NonNullGraphType<AddressType>>>)
+            type: typeof(NonNullGraphType<AddressType>),
+            name: "Signer",
+            description: "An address of the account who signed this transaction.",
+            resolve: x => x.Source.Signer
             );
-            Field(x => x.Signature, type: typeof(NonNullGraphType<ByteStringType>));
-            Field(x => x.Timestamp);
-            Field<NonNullGraphType<ListGraphType<NonNullGraphType<ActionType<T>>>>>("Actions");
+            Field<NonNullGraphType<ByteStringType>>(
+            name: "PublicKey",
+            description: "A PublicKey of the account who signed this transaction.",
+            resolve: ctx => ctx.Source.PublicKey.Format(true)
+            );
+            Field(
+            type: typeof(NonNullGraphType<ListGraphType<NonNullGraphType<AddressType>>>),
+            name: "UpdatedAddresses",
+            description: "Addresses whose states were affected by Actions.",
+            resolve: x => x.Source.UpdatedAddresses
+            );
+            Field(
+            type: typeof(NonNullGraphType<ByteStringType>),
+            name: "Signature",
+            description: "A digital signature of the content of this transaction.",
+            resolve: x => x.Source.Signature
+            );
+            Field(
+            type: typeof(NonNullGraphType<DateTimeOffsetGraphType>),
+            name: "Timestamp",
+            description: "The time this transaction was created and signed.",
+            resolve: x => x.Source.Timestamp
+            );
+            Field<NonNullGraphType<ListGraphType<NonNullGraphType<ActionType<T>>>>>(
+                name: "Actions",
+                description: "A list of actions in this transaction."
+                );
 
             // The block including the transaction. - Only RichStore supports.
             Field<ListGraphType<NonNullGraphType<BlockType<T>>>>(
                 name: "BlockRef",
+                description: "The block including the transaction.",
                 resolve: ctx =>
                 {
                     // FIXME: use context with DI.